Move Pascal frontend code under Pascal directory
