+++ src/Pascal/globals.c
