+++ src/Pascal/globals.h
