+++ src/backend_ast/sdl.c
@@ -16,7 +16,7 @@
 #include "vm/vm.h"
 
 #include "sdl.h" // This header includes SDL/SDL_ttf headers
-#include "globals.h" // Includes SDL.h and SDL_ttf.h via its includes, and audio.h
+#include "Pascal/globals.h" // Includes SDL.h and SDL_ttf.h via its includes, and audio.h
 
 // SDL Global Variable Definitions
 SDL_Window* gSdlWindow = NULL;
