+++ src/symbol/symbol.c
@@ -6,7 +6,7 @@
 //
 #include "Pascal/lexer.h"
 #include "core/utils.h"
-#include "globals.h"
+#include "Pascal/globals.h"
 #include <string.h>
 #include <ctype.h>
 #include <stdbool.h>
